Site: Multiple changes - Truflle network config, setting scripts for contract connect, implementing getFundBalance functions..
So this is a bit of a messy commit but it will do for now. I created a migration file for PHBacon so it can be deployed onto a network. Then deployed it onto a ganache chain and implemented the contract details in the script so it can connect. Then implemented script to update the fundBalance.