SC: One event, two transaction types (err for withdrawl)
We have chosen to emit one Transaction event for both deposits and withdrawl, while presenting the type of withdrawl as an argument of the event (_type bytes32). This arguments holds a byte32 version of the string "deposit" or "withdrawl" which can be used to filter for when colecting the events, and thus can be used to calculate number of deposits or withdrawls, and calculate the total value change of the fund.

However! ========
In remix, when we call the withdrawl function the _type argument correctly showcases the bytes32 version of "withdrawl". However, in truffle when we do so, the _type argument for withdrawl events is similar to the bytes32 of "deposit". This is really weird and needs to be sorted out.